Move serialization helper structs and functions to a separate utils mod
